Dedicated package for scheduler interaction with DRA structured types
