+++ staging/src/k8s.io/dynamic-resource-allocation/structured/schedulerapi/OWNERS
@@ -1,0 +1,6 @@
+# See the OWNERS file documentation: https://git.k8s.io/community/contributors/guide/owners.md
+# Disable inheritance as code changes under this package should be known to sig-autoscaling team.
+options:
+  no_parent_owners: true
+approvers:
+  - sig-autoscaling-maintainers
